Improved context and shlex parsing
