+++ crates/shipit-providers/Cargo.toml
@@ -13,6 +13,7 @@ serde_json.workspace = true
 regex = "1"
 toml = "0.8"
 globset = "0.4"
+shlex = "2"
 walkdir = "2"
 
 [dev-dependencies]
